# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rheo API",
    "description": "Rheo Sovereign AI Voice Studio & Neural Speech Synthesis Engine \u2014 Architected by masterz1311",
    "version": "0.5.0"
  },
  "paths": {
...
    "/generate": {
      "post": {
        "summary": "Generate Speech",
        "description": "Generate speech from text using a voice profile.",
        "operationId": "generate_speech_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/transcribe": {
      "post": {
        "summary": "Transcribe Audio",
        "description": "Transcribe audio file to text.",
        "operationId": "transcribe_audio_transcribe_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_transcribe_audio_transcribe_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/llm/generate": {
      "post": {
        "summary": "Llm Generate",
        "description": "Run a single-turn Qwen3 completion.",
        "operationId": "llm_generate_llm_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_transcribe_audio_transcribe_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_transcribe_audio_transcribe_post"
      },
...
      "EffectConfig": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "EffectConfig",
        "description": "A single effect in an effects chain."
      },
...
      "GenerationRequest": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "text": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 1,
            "title": "Text"
          },
          "language": {
            "type": "string",
            "pattern": "^(zh|en|ja|ko|de|fr|ru|pt|es|it|he|ar|da|el|fi|hi|ms|nl|no|pl|sv|sw|tr)$",
            "title": "Language",
            "default": "en"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "model_size": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(1\\.7B|0\\.6B|1B|3B)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Size",
            "default": "1.7B"
          },
          "instruct": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Instruct"
          },
          "engine": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(qwen|qwen_custom_voice|luxtts|chatterbox|chatterbox_turbo|tada|kokoro)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine",
            "default": "qwen"
          },
          "personality": {
            "type": "boolean",
            "title": "Personality",
            "description": "When true and the profile has a personality prompt, the input text is rewritten in-character before TTS.",
            "default": false
          },
          "max_chunk_chars": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 100.0,
            "title": "Max Chunk Chars",
            "description": "Max characters per chunk for long text splitting",
            "default": 800
          },
          "crossfade_ms": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Crossfade Ms",
            "description": "Crossfade duration in ms between chunks (0 for hard cut)",
            "default": 50
          },
          "normalize": {
            "type": "boolean",
            "title": "Normalize",
            "description": "Normalize output audio volume",
            "default": true
          },
          "effects_chain": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EffectConfig"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effects Chain",
            "description": "Effects chain to apply after generation (overrides profile default)"
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "text"
        ],
        "title": "GenerationRequest",
        "description": "Request model for voice generation."
      },
      "GenerationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "audio_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Path"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "instruct": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instruct"
          },
          "engine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine",
            "default": "qwen"
          },
          "model_size": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Size"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "completed"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "is_favorited": {
            "type": "boolean",
            "title": "Is Favorited",
            "default": false
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "manual"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "versions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/GenerationVersionResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Versions"
          },
          "active_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Version Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "profile_id",
          "text",
          "language",
          "created_at"
        ],
        "title": "GenerationResponse",
        "description": "Response model for voice generation."
      },
...
      "GenerationVersionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "generation_id": {
            "type": "string",
            "title": "Generation Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "audio_path": {
            "type": "string",
            "title": "Audio Path"
          },
          "effects_chain": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EffectConfig"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effects Chain"
          },
          "source_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Version Id"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "generation_id",
          "label",
          "audio_path",
          "is_default",
          "created_at"
        ],
        "title": "GenerationVersionResponse",
        "description": "Response model for a generation version."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMGenerateRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 1,
            "title": "Prompt"
          },
          "system": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "System"
          },
          "model_size": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(0\\.6B|1\\.7B|4B)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Size",
            "default": "0.6B"
          },
          "max_tokens": {
            "type": "integer",
            "maximum": 4096.0,
            "minimum": 1.0,
            "title": "Max Tokens",
            "default": 512
          },
          "temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 0.7
          },
          "examples": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "array",
                "maxItems": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Examples"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "LLMGenerateRequest",
        "description": "Request model for LLM text generation."
      },
      "LLMGenerateResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "model_size": {
            "type": "string",
            "title": "Model Size"
          }
        },
        "type": "object",
        "required": [
          "text",
          "model_size"
        ],
        "title": "LLMGenerateResponse",
        "description": "Response model for LLM text generation."
      },
...
      "TranscriptionResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "duration": {
            "type": "number",
            "title": "Duration"
          }
        },
        "type": "object",
        "required": [
          "text",
          "duration"
        ],
        "title": "TranscriptionResponse",
        "description": "Response model for transcription."
      },
...
    }
  }
}